flaskser/requirements.txt
Flask==3.1.0
requests==2.32.3
Pillow==10.4.0
together==1.3.13